extract useOwner too many errors
